chore: move Nav to dedicated folder
