+++ src/components/ui/Nav/MenuIcon.tsx
@@ -1,16 +1,15 @@
-import { FC } from "react";
-import { IconBase } from "./IconBase";
-
-interface Props {
-  styleName: string;
-}
-
-export const MenuIcon: FC<Props> = ({ styleName }) => {
+const MenuIcon = () => {
   return (
-    <IconBase styleName={styleName} viewBox="0 0 40 40">
+    <svg
+      xmlns="http://www.w3.org/2000/svg"
+      viewBox="0 0 40 40"
+      className="w-10 text-instillGrey05"
+    >
       <line x1="9" y1="14" x2="30" y2="14" stroke="#F7F7F7" strokeWidth="2" />
       <line x1="9" y1="20" x2="30" y2="20" stroke="#F7F7F7" strokeWidth="2" />
       <line x1="15" y1="26" x2="30" y2="26" stroke="#F7F7F7" strokeWidth="2" />
-    </IconBase>
+    </svg>
   );
 };
+
+export default MenuIcon;
